Start refactoring test utils module.
