+++ src/app/test-utils/mock-ngb-active-modal.ts
@@ -1,0 +1,14 @@
+/**
+ * Mock active modal
+ * @packageDocumentation
+ */
+
+/**
+ * Mock active modal
+ */
+export class MockNgbActiveModal {
+  /**
+   * Mock the active modal's close method.
+   */
+  close(): void {}
+}
